requirements.txt
fastapi[standard]
pyrogram
tgcrypto
python-dotenv
python-magic
motor
